statement_id should indeed be usize
StatementID is declared as `int` in the go server code (which would be
equivalent to Rust's `isize`), but based on [the comment][code] with the
declaration, it sounds like StatementID will always be a positive index.

Even if the client is a 32-bit system and the server is a 64-bit system,
since this should correspond to the number of queries sent from the
client, we don't think a 32-bit system would be able to send enough
queries to get a StatementID returned over 32 bits.

[code]:
https://github.com/influxdata/influxdb/blob/933702549da48803df33969a9b7b56476add0cdc/influxql/result.go#L59-L61